requirements.txt
clickclick>=0.14
pystache
PyYAML
dnspython3
stups-pierone>=0.7
boto3>=1.1.3
botocore>=1.2.1
pytest>=2.7.3
